+++ src/Framework/Messages/CommandBus.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Smolblog\Framework;
+namespace Smolblog\Framework\Messages;
 
 /**
  * Object that can pass a Command to its handling service.
